# apps/web/openapi.json
{
  "components": {
    "schemas": {
      "AnswersIn": {
        "additionalProperties": false,
        "properties": {
          "answers": {
            "additionalProperties": {
              "maxLength": 2000,
              "type": "string"
            },
            "description": "open question id -> the employee's answer",
            "maxProperties": 50,
            "title": "Answers",
            "type": "object"
          }
        },
        "required": [
          "answers"
        ],
        "title": "AnswersIn",
        "type": "object"
      },
...
      "ClaimView": {
        "additionalProperties": false,
        "description": "A claim plus where it came from and how it will be routed.",
        "properties": {
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "currency": {
            "default": "INR",
            "title": "Currency",
            "type": "string"
          },
          "document_ids": {
            "items": {
              "type": "string"
            },
            "title": "Document Ids",
            "type": "array"
          },
          "employee_id": {
            "title": "Employee Id",
            "type": "string"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "findings": {
            "description": "Claim-level findings",
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ClaimMode"
          },
          "open_questions": {
            "items": {
              "$ref": "#/components/schemas/OpenQuestion"
            },
            "title": "Open Questions",
            "type": "array"
          },
          "route": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "status": {
            "$ref": "#/components/schemas/ClaimStatus",
            "default": "draft"
          },
          "submission_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Reference"
          },
          "title": {
            "description": "e.g. 'Pune trip 12-14 Aug' or 'Local conveyance Sept 2026'",
            "title": "Title",
            "type": "string"
          },
          "total": {
            "description": "Sum of document totals, in the claim currency",
            "title": "Total",
            "type": "number"
          }
        },
        "required": [
          "id",
          "employee_id",
          "title",
          "mode",
          "document_ids",
          "total"
        ],
        "title": "ClaimView",
        "type": "object"
      },
...
      "DecisionIn": {
        "additionalProperties": false,
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "comment": {
            "default": "",
            "maxLength": 500,
            "title": "Comment",
            "type": "string"
          }
        },
        "required": [
          "approved"
        ],
        "title": "DecisionIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Problem": {
        "description": "The body of every error response (``application/problem+json``).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "PromptOut": {
        "properties": {
          "open_question_ids": {
            "items": {
              "type": "string"
            },
            "title": "Open Question Ids",
            "type": "array"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ONE message with every open question; null if none",
            "title": "Prompt"
          }
        },
        "required": [
          "prompt",
          "open_question_ids"
        ],
        "title": "PromptOut",
        "type": "object"
      },
...
      "ReplyIn": {
        "additionalProperties": false,
        "properties": {
          "text": {
            "description": "Free-text reply to the question",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ReplyIn",
        "type": "object"
      },
...
      "SubmitIn": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "description": "Must be true: the employee explicitly confirms",
            "title": "Confirmed",
            "type": "boolean"
          }
        },
        "required": [
          "confirmed"
        ],
        "title": "SubmitIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI expense & reimbursement agent: from a pile of receipts to ready-to-submit.",
    "title": "ClaimPilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/approvals": {
      "get": {
        "operationId": "approvals_v1_approvals_get",
        "parameters": [
          {
            "description": "submitted | approved | rejected",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "submitted",
              "description": "submitted | approved | rejected",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClaimView"
                  },
                  "title": "Response Approvals V1 Approvals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claims awaiting a decision (approvers)",
        "tags": [
          "claims"
        ]
      }
    },
...
    "/v1/claims": {
      "get": {
        "operationId": "list_claims_v1_claims_get",
        "parameters": [
          {
            "description": "draft|needs_info|ready|submitted|...",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "draft|needs_info|ready|submitted|...",
              "title": "Status"
            }
          },
          {
            "description": "auto_approve | finance_review",
            "in": "query",
            "name": "route",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "auto_approve | finance_review",
              "title": "Route"
            }
          },
          {
            "description": "Approvers only",
            "in": "query",
            "name": "employee_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Approvers only",
              "title": "Employee Id"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClaimView"
                  },
                  "title": "Response List Claims V1 Claims Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claims you may see, newest first",
        "tags": [
          "claims"
        ]
      }
    },
    "/v1/claims/{claim_id}": {
      "get": {
        "operationId": "get_claim_v1_claims__claim_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One claim",
        "tags": [
          "claims"
        ]
      }
    },
    "/v1/claims/{claim_id}/answers": {
      "post": {
        "operationId": "post_answers_v1_claims__claim_id__answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswersIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimView"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Answer one or more open questions",
        "tags": [
          "claims"
        ]
      }
    },
    "/v1/claims/{claim_id}/decision": {
      "post": {
        "operationId": "post_decision_v1_claims__claim_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimView"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve or reject a submitted claim (approvers)",
        "tags": [
          "claims"
        ]
      }
    },
    "/v1/claims/{claim_id}/prompt": {
      "get": {
        "operationId": "get_prompt_v1_claims__claim_id__prompt_get",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The single message that asks everything still open",
        "tags": [
          "claims"
        ]
      }
    },
    "/v1/claims/{claim_id}/reply": {
      "post": {
        "operationId": "post_reply_v1_claims__claim_id__reply_post",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Reply in plain words; the assistant works out which questions it answers",
        "tags": [
          "claims"
        ]
      }
    },
    "/v1/claims/{claim_id}/submit": {
      "post": {
        "operationId": "post_submit_v1_claims__claim_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "description": "Acting persona (employee id)",
            "in": "header",
            "name": "x-persona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Acting persona (employee id)",
              "title": "X-Persona"
            }
          },
          {
            "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
            "in": "header",
            "name": "x-sandbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Demo sandbox id: a visitor's private copy of the demo data (demo only)",
              "title": "X-Sandbox"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimView"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Confirm and submit to the finance system (idempotent per claim)",
        "tags": [
          "claims"
        ]
      }
    },
...
  }
}